aws-s3: remove old implementation
